Move compaction job update event objects to core
