+++ java/core/src/main/java/sleeper/core/tracker/compaction/job/update/CompactionJobFailedEvent.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.core.job.status;
+package sleeper.core.tracker.compaction.job.update;
 
 import sleeper.core.record.process.ProcessRunTime;
